test: Split up unit and integration tests
