+++ daserver/test/integration/__init__.py
@@ -1,0 +1,4 @@
+from django.test import TestCase
+
+
+class DaServerIntegrationTest(TestCase): pass
