+++ daserver/test/unit/__init__.py
@@ -1,0 +1,4 @@
+from django.test import TestCase
+
+
+class DaServerUnitTest(TestCase): pass
