+++ daserver/test/unit/test_package_downloader.py
@@ -1,8 +1,5 @@
 import os
 
-from django.test import TestCase
-
-import common
 from dasdaemon.managers import (
     PathManager,
     QueueManager,
@@ -16,13 +13,15 @@ from dasdaemon.workers import (
 )
 from dasdapi.models import Torrent, PackageFile
 
+import test.common as common
+from test.unit import DaServerUnitTest
 
-class PackageDownloaderTests(TestCase):
+
+class PackageDownloaderUnitTests(DaServerUnitTest):
 
     def setUp(self):
         # Get test config
         self.config = common.load_test_config()
-        #self.num_package_files = self.config['PackageDownloader']['num_package_files']
 
         # Create managers
         self.qm = QueueManager()
@@ -30,9 +29,6 @@ class PackageDownloaderTests(TestCase):
         self.pm = PathManager(config=self.config)
 
         self.package_files_dir = self.pm.package_files_dir.path
-
-        # Create test helper
-        self.helper = common.TestHelper(config=self.config, requests_manager=self.rm)
 
         # Create instance
         self.pd = PackageDownloader(
@@ -50,7 +46,7 @@ class PackageDownloaderTests(TestCase):
         utils.fs.rm_rf(self.package_files_dir)
 
     def _set_no_write_dir(self, dirpath):
-        os.chmod(dirpath, 0555)
+        os.chmod(dirpath, 0o555)
 
     def _create_package_files(self, package_files_count, torrent_name='Torrent'):
         # Add Torrent to database
@@ -121,38 +117,6 @@ class PackageDownloaderTests(TestCase):
         t = Torrent.objects.get(pk=1)
         self.assertEqual(t.stage, PackageDownloader.processing_stage())
 
-    def test_do_work_one_file(self):
-        # Create package file on remote host
-        remote_filename = 'Packaged.file'
-        r = self.helper.create_packaged_torrent(remote_filename)
-        self.assertEqual(r.status_code, 200)
-
-        # Add package file to database
-        torrent = Torrent.objects.create(name='Torrent')
-        package_file = PackageFile.objects.create(
-            filename='%s.0000' % remote_filename,
-            torrent=torrent,
-            stage=self.pd.package_file_ready_stage()
-        )
-
-        # Register as consumer
-        self.pd.register_as_consumer()
-
-        # Run queue manager
-        self.qm._execute_queries()
-
-        # Run package extractor
-        self.pd.do_work()
-
-        # Verify package file updated
-        package_file.refresh_from_db()
-        self.assertEqual(self.pd.package_file_completed_stage(), package_file.stage)
-
-        # Verify file exists
-        package_file_path = os.path.join(self.package_files_dir, package_file.filename)
-        package_file_path = self.pm.get_package_file_path(torrent, package_file)
-        self.assertTrue(os.path.isfile(package_file_path))
-
     def test_do_work_nonexistant_file(self):
         # Add package file to database
         torrent = Torrent.objects.create(name='Torrent')
@@ -168,7 +132,7 @@ class PackageDownloaderTests(TestCase):
         # Run queue manager
         self.qm._execute_queries()
 
-        # Run package extractor
+        # Run package downloader
         self.pd.do_work()
 
         # Verify file does not exist
@@ -179,85 +143,3 @@ class PackageDownloaderTests(TestCase):
         # Verify package file is in error stage
         self.assertEqual('Error', package_file.stage)
         self.assertEqual(1, package_file.errors.count())
-
-    def test_do_work_no_perms(self):
-        # Create Package File on remote host
-        remote_filename = 'Packaged.file'
-        r = self.helper.create_packaged_torrent(remote_filename)
-        self.assertEqual(r.status_code, 200)
-
-        # Add Package File to database
-        torrent = Torrent.objects.create(name='Torrent')
-        package_file = PackageFile.objects.create(
-            filename='%s.0000' % remote_filename,
-            torrent=torrent,
-            stage=self.pd.package_file_ready_stage()
-        )
-
-        # Set directory read only
-        self._set_no_write_dir(self.package_files_dir)
-
-        # Register as consumer
-        self.pd.register_as_consumer()
-
-        # Run queue manager
-        self.qm._execute_queries()
-
-        # Run package extractor
-        self.pd.do_work()
-
-        # Verify file does not exist
-        package_file.refresh_from_db()
-        package_file_path = self.pm.get_package_file_path(torrent, package_file)
-        self.assertFalse(os.path.isfile(package_file_path))
-
-        # Verify package file is in error stage
-        self.assertEqual('Error', package_file.stage)
-        self.assertEqual(1, package_file.errors.count())
-
-    def test_do_work_two_files(self):
-        # Create package file on remote host
-        remote_filename = 'Packaged.file'
-        r = self.helper.create_packaged_torrent(remote_filename, file_count=2)
-        self.assertEqual(r.status_code, 200)
-
-        # Add package files to database
-        torrent = Torrent.objects.create(name='Torrent')
-        package_file1 = PackageFile.objects.create(
-            filename='%s.0000' % remote_filename,
-            torrent=torrent,
-            stage=self.pd.package_file_ready_stage()
-        )
-        package_file2 = PackageFile.objects.create(
-            filename='%s.0001' % remote_filename,
-            torrent=torrent,
-            stage=self.pd.package_file_ready_stage()
-        )
-
-        # Register as consumer
-        self.pd.register_as_consumer()
-
-        # Run queue manager
-        self.qm._execute_queries()
-
-        # Run package extractor
-        self.pd.do_work()
-
-        # Verify package file updated
-        package_file1.refresh_from_db()
-        self.assertEqual(self.pd.package_file_completed_stage(), package_file1.stage)
-
-        # Verify files exist
-        package_file_path1 = self.pm.get_package_file_path(torrent, package_file1)
-        self.assertTrue(os.path.isfile(package_file_path1))
-
-        # Run package extractor again
-        self.pd.do_work()
-
-        # Verify package file updated
-        package_file2.refresh_from_db()
-        self.assertEqual(self.pd.package_file_completed_stage(), package_file2.stage)
-
-        # Verify files exist
-        package_file_path2 = self.pm.get_package_file_path(torrent, package_file2)
-        self.assertTrue(os.path.isfile(package_file_path2))
